rename stat query group and sort
