+++ java/settings.gradle
@@ -83,7 +83,7 @@ include ':add-graphics-renderer',
         ':spatial-relationships',
         ':surface-placement',
         ':statistical-query',
-        ':statistical-query-group-sort',
+        ':statistical-query-group-and-sort',
         ':stretch-renderer',
         ':symbolize-shapefile',
         ':take-screenshot',
