+++ java/statistical-query-group-and-sort/src/main/java/com/esri/arcgisruntime/sample/statisticalquerygroupandsort/MainActivity.java
@@ -14,7 +14,7 @@
  *
  */
 
-package com.esri.arcgisruntime.sample.statisticalquery;
+package com.esri.arcgisruntime.sample.statisticalquerygroupandsort;
 
 import java.util.ArrayList;
 import java.util.Iterator;
