+++ java/statistical-query-group-and-sort/src/main/res/drawable-v24/ic_launcher_foreground.xml
