+++ java/statistical-query-group-and-sort/src/main/res/drawable/ic_launcher_background.xml
