+++ java/statistical-query-group-and-sort/src/main/res/layout/field_checkbox_row.xml
