+++ java/statistical-query-group-and-sort/src/main/res/layout/group_fields_view.xml
