+++ java/statistical-query-group-and-sort/src/main/res/layout/query_executing_dialog.xml
