+++ java/statistical-query-group-and-sort/src/main/res/layout/results_expandablelistview.xml
