+++ java/statistical-query-group-and-sort/src/main/res/layout/statistics_view.xml
