+++ java/statistical-query-group-and-sort/src/main/res/mipmap-hdpi/ic_launcher_round.png
