+++ java/statistical-query-group-and-sort/src/main/res/mipmap-mdpi/ic_launcher.png
