+++ java/statistical-query-group-and-sort/src/main/res/mipmap-xhdpi/ic_launcher.png
