+++ java/statistical-query-group-and-sort/src/main/res/mipmap-xxhdpi/ic_launcher_round.png
