+++ java/statistical-query-group-and-sort/src/main/res/mipmap-xxxhdpi/ic_launcher_round.png
