+++ java/statistical-query-group-and-sort/src/main/res/values/strings.xml
